+++ sdk/core/azure_core/src/http/policies/mod.rs
@@ -3,12 +3,11 @@
 
 //! HTTP pipeline policies.
 
-mod bearer_token_policy;
+pub mod auth;
 mod client_request_id;
 mod instrumentation;
 mod user_agent;
 
-pub use bearer_token_policy::BearerTokenAuthorizationPolicy;
 pub use client_request_id::*;
 pub use instrumentation::*;
 pub use typespec_client_core::http::policies::*;